refactor(commands): make ls a top level subcommand
- simplify calling ergonomics
- improve subcommand organization for users, groups, etc.
