+++ compiler_admin/commands/ls/users.py
@@ -1,0 +1,72 @@
+import click
+import pandas as pd
+
+from compiler_admin import FORMATS, Format
+from compiler_admin.services import files
+from compiler_admin.services.google import get_users
+from compiler_admin.services.toggl import TogglUsers
+
+
+def google(format: int = Format.BASIC, inactive: bool = False, **kwargs):
+    """Use GAM to print the users in the Google Workspace."""
+    output = get_users(inactive=inactive, format=format, **kwargs)
+    click.echo(output)
+
+
+def toggl(format: int = Format.BASIC, inactive: bool = False, **kwargs):
+    """Use the Toggl API to get a list of users.
+
+    Mirrors the GAM output style for Google users.
+    """
+    api = TogglUsers()
+
+    click.echo("Getting all Toggl users...", err=True)
+    users = api.get_organization_users(inactive=inactive, **kwargs)
+    users_df = pd.DataFrame(users)
+    click.echo(f"Got {len(users)} Users", err=True)
+
+    stdout = click.get_text_stream("stdout")
+    columns = ["email"]
+    if format == Format.BASIC:
+        files.write_csv(stdout, users_df, columns=columns)
+    elif format == Format.CSV:
+        columns += [
+            "name",
+            "id",
+            "user_id",
+            "role_id",
+            "organization_id",
+            "inactive",
+            "joined",
+            "can_edit_email",
+            "2fa_enabled",
+            "avatar_url",
+        ]
+        files.write_csv(stdout, users_df, columns=columns)
+    elif format == Format.JSON:
+        files.write_json(stdout, users)
+
+
+USER_SYSTEMS = {"google": google, "toggl": toggl}
+
+
+@click.command()
+@click.option(
+    "--format",
+    "format_key",
+    help="The format of the output.",
+    type=click.Choice(FORMATS.keys(), case_sensitive=False),
+    default="basic",
+)
+@click.option("--inactive", help="List inactive users in the system.", is_flag=True)
+@click.argument(
+    "system",
+    type=click.Choice(USER_SYSTEMS.keys(), case_sensitive=False),
+    default="google",
+)
+def users(system: str, inactive: bool, format_key: str, **kwargs):
+    """List users in the Compiler workspace."""
+    format = FORMATS.get(format_key)
+
+    ls_system = USER_SYSTEMS.get(system)
+    ls_system(format=format, inactive=inactive, **kwargs)
