+++ docs/reference/cli/ls.md
@@ -1,0 +1,9 @@
+# `compiler-admin ls`
+
+<!-- prettier-ignore -->
+::: mkdocs-click
+    :module: compiler_admin.main
+    :command: ls
+    :depth: 1
+    :prog_name: compiler-admin ls
+    :list_subcommands: True
