+++ tests/commands/user/test_init.py
@@ -11,7 +11,6 @@ from compiler_admin.commands.user import user
         "create",
         "deactivate",
         "delete",
-        "ls",
         "offboard",
         "reactivate",
         "reset",
