+++ tests/test_main.py
@@ -3,6 +3,6 @@ import pytest
 from compiler_admin.main import main
 
 
-@pytest.mark.parametrize("command", ["info", "init", "time", "user"])
+@pytest.mark.parametrize("command", ["info", "init", "ls", "time", "user"])
 def test_main_commands(command):
     assert command in main.commands
